# Improve performance of obj file loading